Tweak so that it generates output (problematic?)
I made this change because the model was erroneously using a different feature as the 'label' when running on the com2sense set, causing it not to generate output. This change made it generate output, but it involved a modification outside of the todo block and may still not produce correct behavior.

Another possible solution would be not to make this change and to train and evaluate using different commands. I couldn't figure out how to do this and was running low on time, but in theory args.do_train being false would result in the same behavior. (The given shell script doesn't do this, though - so maybe we have some bug in data loading that is causing this issue?)